Move HookEventBridge into Friendica\Core\Hooks namespace
